Refactor Algortihm and add JSON and GeoJSON support
