+++ opentabulate/main/algorithms/__init__.py
@@ -1,0 +1,5 @@
+from .algorithm         import Algorithm
+from .csv_algorithm     import CSV_Algorithm
+from .geojson_algorithm import GeoJSON_Algorithm
+from .json_algorithm    import JSON_Algorithm
+from .xml_algorithm     import XML_Algorithm
